zcash_client_backend: Move `SaplingPoolTester` here from `zcash_client_sqlite`
