+++ zcash_client_sqlite/src/chain.rs
@@ -322,7 +322,9 @@ where
 #[cfg(test)]
 #[allow(deprecated)]
 mod tests {
-    use crate::{testing, wallet::sapling::tests::SaplingPoolTester};
+    use zcash_client_backend::data_api::testing::sapling::SaplingPoolTester;
+
+    use crate::testing;
 
     #[cfg(feature = "orchard")]
     use crate::wallet::orchard::tests::OrchardPoolTester;
